Move BootstrapProxy to its own top-level class
