+++ dd-java-agent/testing/src/main/groovy/datadog/trace/agent/test/SpockRunner.java
@@ -2,7 +2,7 @@ package datadog.trace.agent.test;
 
 import com.google.common.reflect.ClassPath;
 import datadog.trace.agent.test.utils.ClasspathUtils;
-import datadog.trace.bootstrap.DatadogClassLoader.BootstrapClassLoaderProxy;
+import datadog.trace.bootstrap.BootstrapProxy;
 import java.io.File;
 import java.io.IOException;
 import java.lang.reflect.Field;
@@ -160,7 +160,7 @@ public class SpockRunner extends Sputnik {
           .appendToBootstrapClassLoaderSearch(new JarFile(bootstrapJar));
       // Utils cannot be referenced before this line, as its static initializers load bootstrap
       // classes (for example, the bootstrap proxy).
-      ((BootstrapClassLoaderProxy) datadog.trace.agent.tooling.Utils.getBootstrapProxy())
+      ((BootstrapProxy) datadog.trace.agent.tooling.Utils.getBootstrapProxy())
           .addURL(bootstrapJar.toURI().toURL());
     } catch (final IOException e) {
       throw new RuntimeException(e);
